Initialize React and NodeJS PWA App
Use create-react-app to create a template PWA app.
Create a script to get viewport size and implement a message stating that
this app is only available on mobile. If the app is open in the browser,
it should display a message instructing the participant to add the app
to their homescreen.